Move tests under /python/ray/tests; Updated BUILD file
